Moving mailiosmtp object to separate module
