+++ types/mailiosmtp/go.mod
@@ -1,0 +1,3 @@
+module github.com/mailio/go-mailio-server/types/mailiosmtp
+
+go 1.22.1
